mouse_emulator/pyproject.toml
[project]
name = "mouse-emulator"
version = "0.1.0"
description = "Keyboard driven mouse automation for macOS"
requires-python = ">=3.11"
readme = "README.md"
authors = [{ name = "mouse_emulator_team" }]
dependencies = [
  "pynput>=1.7.6",
  "typer>=0.12.3",
  "pydantic>=2.6.4",
  "tabulate>=0.9.0",
  "pyyaml>=6.0.2",
  "numpy>=1.26.4",
  "opencv-python-headless>=4.10.0.84",
  "pytesseract>=0.3.13",
  "pillow>=10.4.0",
]

[dependency-groups]
dev = [
  "mypy>=1.18.2",
  "pytest>=8.4.2",
  "ruff>=0.14.2",
  "types-pynput>=0.0.9",
  "types-tabulate>=0.9.0.20240106",
  "types-PyYAML>=6.0.12.20240917",
  "types-Pillow>=10.2.0.20240822",
]

[project.scripts]
mouse-emulator = "mouse_emulator.cli:app"
auto-emulator = "auto_emulator.cli:app"

[tool.uv]
package = true

[tool.ruff]
line-length = 88
target-version = "py311"
preview = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "A001",
  "C408",
  "C901",
  "COM812",
  "CPY001",
  "D100",
  "D101",
  "D102",
  "D103",
  "D105",
  "D107",
  "D415",
  "EM101",
  "EM102",
  "FBT001",
  "PLR0911",
  "PLR0914",
  "PLR0915",
  "PLR2004",
  "PLR6301",
  "T201",
  "TC001",
  "TC002",
  "TC003",
  "TRY003",
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["mouse_emulator", "mouse_core", "auto_emulator"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
